test(linter/plugins): fix garbled output from assertions (#16738)
Conformance tester has been suffering from non-deterministic output from NodeJS's `assert` module. Fix this by setting `process.env.FORCE_COLOR` before loading any code.

This also allows leaving the `     ^` diff marker lines in output which were removed in #16723, as they should now appear consistently.
